+++ packages/flrx_test/lib/flrx_test.dart
@@ -1,0 +1,4 @@
+library flrx_test;
+
+export 'package:flrx_test/src/application.dart';
+export 'package:flrx_test/src/store.dart';