+++ packages/framework/lib/components/error_manager/error_manager.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/foundation.dart';
+
+export 'catcher_error_manager.dart';
+
+abstract class ErrorManager {
+  ErrorManager();
+
+  void init(VoidCallback runAppFunction);
+
+  void reportCheckedError(dynamic error, dynamic stackTrace);
+}