+++ packages/framework/lib/pages/page.dart
@@ -16,7 +16,6 @@ import 'package:redux/redux.dart';
 /// a [ViewModel] defined for the widget using this mixin.
 ///
 abstract class Page<State, V extends ViewModel<State>> {
-
   /// A function that will run when the [StoreConnector] in [Page] will
   /// initially be created.
   ///