+++ packages/framework/lib/pages/page_not_found.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 class PageNotFound extends StatelessWidget {
-
   const PageNotFound({Key? key}) : super(key: key);
 
   @override