+++ packages/framework/test/unit/application_test.dart
@@ -10,7 +10,7 @@ void main() {
   test('That Application can initialize Modules', () {
     var mockModule = MockModule();
 
-    Application.init(() async {}, config: MockConfig([mockModule]));
+    Application(() async {}, appConfig: MockConfig([mockModule])).init();
 
     var instance = Application.get<RandomClass>();
     expect(instance is RandomClass, true);